requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/requirements.txt requirements/requirements.in
#
asgiref==3.2.3            # via django
attrs==19.3.0
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via clickclick, flask
clickclick==1.2.2         # via connexion
connexion[swagger-ui]==2.4.0
coverage==4.5.4
dependencies==0.15
django-db-prefix==1.0.5
django-extensions==2.2.5
django-guardian==2.1.0
django==3.0
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==3.0.0              # via factory-boy
filelock==3.0.12          # via tox
filemagic==1.6
flake8-coding==1.3.2
flake8-comprehensions==3.1.4
flake8-debugger==3.2.1
flake8-django==0.0.4
flake8-docstrings==1.5.0
flake8==3.7.9
flask-testing==0.7.1
flask==1.1.1
idna==2.8                 # via requests
importlib-metadata==1.2.0  # via flake8-comprehensions, jsonschema, pluggy, pytest, tox
inflection==0.3.1         # via connexion, pytest-factoryboy
isort==4.3.21
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask, swagger-ui-bundle
jsonschema==3.2.0         # via connexion, openapi-spec-validator
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mock==3.0.5
more-itertools==8.0.2     # via pytest, zipp
openapi-spec-validator==0.2.8  # via connexion
packaging==19.2           # via pytest, tox
pluggy==0.13.1            # via pytest, tox
psycopg2-binary==2.8.4
py-spy==0.3.1
py==1.8.0                 # via pytest, tox
pycodestyle==2.5.0        # via flake8, flake8-debugger
pydocstyle==2.1.1
pydotplus==2.0.2
pyflakes==2.1.1           # via flake8
pyjwt==1.7.1
pyparsing==2.4.5          # via packaging, pydotplus
pyrsistent==0.15.6        # via jsonschema
pytest-cov==2.8.1
pytest-django==3.7.0
pytest-factoryboy==2.0.3
pytest-mock==1.13.0
pytest==5.3.1
python-dateutil==2.8.1    # via faker
pytz==2019.3              # via django
pyyaml==5.2               # via clickclick, connexion, openapi-spec-validator
requests==2.22.0          # via connexion
six==1.13.0               # via connexion, django-extensions, faker, jsonschema, mock, openapi-spec-validator, packaging, pydocstyle, pyrsistent, python-dateutil, tox
snowballstemmer==2.0.0    # via pydocstyle
sqlparse==0.3.0           # via django
stories==0.10.1
swagger-ui-bundle==0.0.6  # via connexion
text-unidecode==1.3       # via faker
toml==0.10.0              # via tox
tox==3.14.2
urllib3==1.25.7           # via requests
virtualenv==16.7.8        # via tox
wcwidth==0.1.7            # via pytest
werkzeug==0.16.0          # via flask
zipp==0.6.0               # via importlib-metadata
